chore: rebuild docs from public
